requirement.txt
Flask==3.0.3
openai==1.51.1
pymongo==4.10.1
python-dotenv==1.0.1
flask_cors==5.0.0
uwsgi==2.0.27